[memory_leak] Fixing memory leak in ResourceManager.
ResourceManager is a class used to manage the resource among different tests. A given class may use a ResourceManager to store the resources it produces (such as timers, etc.), and each test can later use the test.util.ModuleDisposer.uses_module decorator to check that all the resources created during the test will be cleaned.

The problem of ResourceManager is that when run in production, it will store forever those resources. In the case of the UserProcessingServer, it will store a new timer every 3 seconds (depending on the configuration of the 'core_checking_time' property). Since nobody removes those objects, they are stored there forever, as well as all the objects created by the Timer (it includes a RLock, a Lock, a Condition, an Event, a dict per each object, etc. etc.), causing the memory to increase steadily.

Performing a:

  del weblab.user_processing.UserProcessingServer._resources_manager._resources[:]

caused the number of objects to decrease from more than 7,000,000 to 175,000, after running the system for almost a month.

With this commit, those resources are not added unless the module "test" is loaded (which is only loaded when running tests).