Support sqlalchemy 0.6. Big performance problems in unit tests. See log detail.
Prior to this commit, we used shared pools for the connections across the unit
tests. In order to do so, pools were defined at class level, and engines (which
are bound to configurable credentials) at instance level. However, since
sqlalchemy 0.6, pools can't be shared across engines (even with the same
configuration), so we now dispose them and recreate them each time. This brings
a performance impact, taking 599 seconds instead of 298 seconds to run the
tests.