Add first version of the reservation_processor. See detailed description.
The idea is to make reservations independent of particular user sessions. Previously, a
user would start a session, and would perform activities with this session, such as
listing experiments, reserving experiments, and working on these reservations (seeing
the status of these reservations, interacting with the experiment, finishing it). The
problem is that working on these reservations is something that should be shareable in
two senses:

 - First, a user might want to delegate other user to participate in a particular
   reservation, without doing anything else in the name of the original user. This is
   the case for LMSs or federated users.

 - Secondly, a user might want to interact with a experiment with other user. It would
   be interesting to share this reservation id and this other user would be able to
   directly interact with the experiment, too.

Therefore, the reservation_processor module has been created. Users will use the
session_id (and therefore, the processor.py module) for regular operations. However,
they will start using the reservation_id (and therefore, the reservation_processor.py
module) once the reservation has been made. Any holder of a reservation_id will be
able to see the scheduling status of the reservation, cancel it, finish the experiment,
or interact with it, but he will not be able to create new reservations neither of
acting in the name of the original user.