Separate the lab client from the admin client.
This was indeed a bug, which we start noticing when the admin client was not
being loaded. However, fixing it has an important impact on the size of the
generated files (around half of the original size). Additionally, the build.xml
now removes the .junit_symbolMap directories, so in total what used to be 200 MB
is now around 50 MB!